+++ sip_router/parser/parse_via.c
@@ -1,5 +1,5 @@
 /*
- * $Id: parse_via.c,v 1.7 2002/09/19 12:23:55 jku Rel $ 
+ * $Id: parse_via.c,v 1.8 2002/11/26 15:29:51 andrei Exp $ 
  *
  * via parsing automaton
  * 
@@ -1794,11 +1794,11 @@ nextvia:
 
 error:
 	if (end>buffer){
-		LOG(L_ERR, "ERROR: parse_via on: <%.*s>\n", end-buffer, buffer);
+		LOG(L_ERR, "ERROR: parse_via on: <%.*s>\n", (int)(end-buffer), buffer);
 	}
 	if ((tmp>buffer)&&(tmp<end)){
 		LOG(L_ERR, "ERROR: parse_via parse error, parsed so far:<%.*s>\n",
-				tmp-buffer, buffer );
+				(int)(tmp-buffer), buffer );
 	}else{
 		LOG(L_ERR, "ERROR: parse_via: via parse error\n");
 	}
